docs: keep the docs sidebar stable across navigation
Move the docs header and sidebar into a persistent app/docs/layout.tsx so
they no longer re-mount on every navigation. The sidebar is now a client
component that derives the active item from usePathname(), which preserves
its scroll position and stops the visible jump when clicking links.

Also reserve the scrollbar gutter (scrollbar-gutter: stable) to prevent
horizontal layout shift between pages of different heights, and switch
scroll-behavior from smooth to auto so scroll-to-top is instantaneous.
