+++ site/app/globals.css
@@ -141,8 +141,9 @@
 	}
 
 	html {
-		scroll-behavior: smooth;
+		scroll-behavior: auto;
 		scroll-padding-top: 5.5rem;
+		scrollbar-gutter: stable;
 	}
 
 	body {
